// galiza-frontend/src/app/app-front/models/enums.ts
export enum Rol {
  USER = 'USER',
  ADMIN = 'ADMIN',
  SUPERUSER = 'SUPERUSER',
}

export enum TipoRecomendacion {
  LIBRO = 'LIBRO',
  ARTISTA = 'ARTISTA',
  CANCION = 'CANCION',
  DOCUMENTAL = 'DOCUMENTAL',
  WEB = 'WEB',
  OTRO = 'OTRO',
}

export enum EstadoSolicitud {
  PENDIENTE = 'PENDIENTE',
  APROBADA = 'APROBADA',
  RECHAZADA = 'RECHAZADA',
}


export enum TipoEvento {
  ROMERIA = 'Romería',
  FESTIVAL = 'Festival',
  TALLER = 'Taller',
  CONCIERTO = 'Concierto',
  CHARLA = 'Charla',
  FOLIADA = 'Foliada',
  PROBA = 'Proba',
  OUTRO = 'Outro'
}


export enum TipoPunto {
  MUINEIRA = 'Muiñeira',
  XOTA = 'Xota',
  VOLTA = 'Volta',
  TABLON = 'Tablón',
  MANEO = 'Maneo',
  PASODOBRE = 'Pasodobre',
  PUNTO_GALAICO = 'Punto Galaico',
  OUTRO = 'Outro'
}

export enum Compas {
  XOTA = 'Xota',
  MUINEIRA_NOVA = 'Muiñeira Nova',
  MUINEIRA_VELLA = 'Muiñeira Vella',
  CARBALLESA = 'Carballesa',
  MANEO = 'Maneo',
  PASODOBRE = 'Pasodobre',
  VALSE = 'Valse',
  POLCA = 'Polca',
  MAZURCA = 'Mazurca',
  RUMBA = 'Rumba',
  PASARRUAS = 'Pasarruas'
}